Moved functions to plot_utils file
